requirements.txt
python-dotenv
google-generativeai
python-docx
tweepy